Revert "Fix bug with pipe color (#30645)"
This reverts commit 41a175636b115b79b551d142c5c97d7d7571fde9.
