+++ Content.Client/UserInterface/Systems/Storage/Controls/ItemGridPiece.cs
@@ -185,12 +185,7 @@ public sealed class ItemGridPiece : Control, IEntityControl
 
             handle.SetTransform(pos, iconRotation);
             var box = new UIBox2(root, root + sprite.Size * scale);
-
-            var ev = new BeforeRenderInGridEvent(new Color(255, 255, 255));
-            _entityManager.EventBus.RaiseLocalEvent(Entity, ev);
-
-            handle.DrawTextureRect(sprite, box, ev.Color);
-
+            handle.DrawTextureRect(sprite, box);
             handle.SetTransform(GlobalPixelPosition, Angle.Zero);
         }
         else
@@ -301,19 +296,6 @@ public sealed class ItemGridPiece : Control, IEntityControl
     }
 
     public EntityUid? UiEntity => Entity;
-}
-
-/// <summary>
-///     This event gets raised before a sprite gets drawn in a grid and lets to change the sprite color for several gameobjects that have special sprites to render in containers.
-/// </summary>
-public sealed class BeforeRenderInGridEvent : EntityEventArgs
-{
-    public Color Color { get; set; }
-
-    public BeforeRenderInGridEvent(Color color)
-    {
-        Color = color;
-    }
 }
 
 public enum ItemGridPieceMarks
