+++ Content.Server/Atmos/Piping/Components/AtmosPipeColorComponent.cs
@@ -1,22 +1,19 @@
-using Content.Shared.Atmos.EntitySystems;
-using Robust.Shared.GameStates;
+using Content.Server.Atmos.Piping.EntitySystems;
 using JetBrains.Annotations;
 
-namespace Content.Shared.Atmos.Components;
+namespace Content.Server.Atmos.Piping.Components;
 
-[RegisterComponent, NetworkedComponent]
-[AutoGenerateComponentState]
+[RegisterComponent]
 public sealed partial class AtmosPipeColorComponent : Component
 {
     [DataField]
-    [AutoNetworkedField]
     public Color Color { get; set; } = Color.White;
 
     [ViewVariables(VVAccess.ReadWrite), UsedImplicitly]
     public Color ColorVV
     {
         get => Color;
-        set => IoCManager.Resolve<IEntityManager>().System<AtmosPipeColorSystem>().SetColor((Owner, this), value);
+        set => IoCManager.Resolve<IEntityManager>().System<AtmosPipeColorSystem>().SetColor(Owner, this, value);
     }
 }
